lambda/get_pricing_cmc/requirements.txt
requests==2.31.0
mysql-connector-python==8.0.32
